responsive web design update , bug fixes
made website suitable for vertical (portrait) view and add experimental swipe controlsfor touch screens.

fixed an infinite loop in function generateFood when the snake fill the grid.

there is a bug that happens in chrome sometimes (because of slow DOM manipulation) so I added a throw error statment *needs fix*